home-astro: extract landing components and modularize animation scripts
- Extract homepage HTML into BootScreen, SplashScreen, TunnelStage, TweaksPanel, and VaporOutro components

- Decompose home.js into ES6 modules under src/scripts/landing (boot.js, centerline.js, tweaks.js)

- Wire up the modules inside index.astro and home.js

- Verify correct page load and boot sequence execution via browser automation
