+++ src/components/landing/BootScreen.astro
@@ -1,0 +1,8 @@
+---
+---
+<div class="layer" id="boot">
+  <div class="boot-content">
+    <div class="head"><span class="prompt"><b>alan@barzilay</b>:~/sites/personal $</span> <span class="cmd">./boot --landing</span></div>
+    <pre id="bootLog"></pre>
+  </div>
+</div>
