+++ src/components/landing/SplashScreen.astro
@@ -1,0 +1,48 @@
+---
+---
+<div class="layer" id="splash">
+  <div class="splash-inner" id="splashInner">
+    <div class="logo-wrap">
+      <!-- left trace -->
+      <div class="logo-half left" id="logoLeft">
+        <svg viewBox="0 0 96.520073 42.426407" aria-hidden="true">
+          <g transform="translate(13.101604,-35.573723)">
+            <path d="M -7.9846602,67.484737 H 10.778432 c 0.3086,0 0.6001,-0.1426 0.7896,-0.3862 l 7.460031,-9.627573 h 14.680346 c 0.4473,1.7202 1.9995,3 3.8579,3 2.2056,0 4,-1.7944 4,-4 0,-2.2056 -1.7944,-4 -4,-4 -1.8584,0 -3.4106,1.2798 -3.8579,3 H 18.538763 c -0.3086,0 -0.6001,0.1426 -0.789601,0.3862 l -7.46003,9.627573 H -7.9846602 c -0.5522,0 -1,0.4478 -1,1 0,0.5522 0.4478,1 1,1 z M 37.566309,54.470964 c 1.103,0 2,0.897 2,2 0,1.103 -0.897,2 -2,2 -1.103,0 -2,-0.897 -2,-2 0,-1.103 0.897,-2 2,-2 z"/>
+          </g>
+        </svg>
+      </div>
+      <!-- right trace -->
+      <div class="logo-half right" id="logoRight">
+        <svg viewBox="0 0 96.520073 42.426407" aria-hidden="true">
+          <g transform="translate(13.101604,-35.573723)">
+            <path d="M 78.301524,57.322218 H 62.709556 l -7.25496,-14.126192 c -0.1895,-0.2436 -0.481,-0.386141 -0.7896,-0.3862 l -26.506279,-0.005 c -0.4473,-1.7202 -1.9995,-3 -3.8579,-3 -2.2056,0 -4,1.7944 -4,4 0,2.2061 1.7944,4 4,4 1.8584,0 3.4106,-1.2795 3.8579,-3 l 26.016979,0.005 5.99956,12.512392 h -8.873732 c -0.3262,0 -0.6323,0.1592 -0.8193,0.4268 L 42.795361,68.75591 27.954427,68.76903 c -0.4473,-1.7205 -1.9995,-3 -3.8579,-3 -2.2056,0 -4,1.7939 -4,4 0,2.2061 1.7944,4 4,4 1.8584,0 3.4106,-1.2795 3.8579,-3 l 15.361434,-0.01312 c 0.3262,-2.79e-4 0.6323,-0.1592 0.8193,-0.4268 l 7.686863,-11.006892 h 26.4795 c 0.5522,0 1,-0.4473 1,-1 0,-0.5527 -0.4478,-1 -1,-1 z M 24.300817,45.804787 c -1.103,0 -2,-0.897 -2,-2 0,-1.103 0.897,-2 2,-2 1.103,0 2,0.897 2,2 0,1.103 -0.897,2 -2,2 z m -0.204292,25.964246 c -1.103,0 -2,-0.8975 -2,-2 0,-1.1025 0.897,-2 2,-2 1.103,0 2,0.8975 2,2 0,1.1025 -0.897,2 -2,2 z"/>
+          </g>
+        </svg>
+      </div>
+    </div>
+    <div class="splash-name" id="splashName">alan barzilay<span class="dot">.</span></div>
+    <div class="splash-sub" id="splashSub">developer · researcher · maker</div>
+  </div>
+  <div class="splash-inner splash-glitch" id="splashInnerGlitch" aria-hidden="true">
+    <div class="logo-wrap">
+      <!-- left trace -->
+      <div class="logo-half left">
+        <svg viewBox="0 0 96.520073 42.426407" aria-hidden="true">
+          <g transform="translate(13.101604,-35.573723)">
+            <path d="M -7.9846602,67.484737 H 10.778432 c 0.3086,0 0.6001,-0.1426 0.7896,-0.3862 l 7.460031,-9.627573 h 14.680346 c 0.4473,1.7202 1.9995,3 3.8579,3 2.2056,0 4,-1.7944 4,-4 0,-2.2056 -1.7944,-4 -4,-4 -1.8584,0 -3.4106,1.2798 -3.8579,3 H 18.538763 c -0.3086,0 -0.6001,0.1426 -0.789601,0.3862 l -7.46003,9.627573 H -7.9846602 c -0.5522,0 -1,0.4478 -1,1 0,0.5522 0.4478,1 1,1 z M 37.566309,54.470964 c 1.103,0 2,0.897 2,2 0,1.103 -0.897,2 -2,2 -1.103,0 -2,-0.897 -2,-2 0,-1.103 0.897,-2 2,-2 z"/>
+          </g>
+        </svg>
+      </div>
+      <!-- right trace -->
+      <div class="logo-half right">
+        <svg viewBox="0 0 96.520073 42.426407" aria-hidden="true">
+          <g transform="translate(13.101604,-35.573723)">
+            <path d="M 78.301524,57.322218 H 62.709556 l -7.25496,-14.126192 c -0.1895,-0.2436 -0.481,-0.386141 -0.7896,-0.3862 l -26.506279,-0.005 c -0.4473,-1.7202 -1.9995,-3 -3.8579,-3 -2.2056,0 -4,1.7944 -4,4 0,2.2061 1.7944,4 4,4 1.8584,0 3.4106,-1.2795 3.8579,-3 l 26.016979,0.005 5.99956,12.512392 h -8.873732 c -0.3262,0 -0.6323,0.1592 -0.8193,0.4268 L 42.795361,68.75591 27.954427,68.76903 c -0.4473,-1.7205 -1.9995,-3 -3.8579,-3 -2.2056,0 -4,1.7939 -4,4 0,2.2061 1.7944,4 4,4 1.8584,0 3.4106,-1.2795 3.8579,-3 l 15.361434,-0.01312 c 0.3262,-2.79e-4 0.6323,-0.1592 0.8193,-0.4268 l 7.686863,-11.006892 h 26.4795 c 0.5522,0 1,-0.4473 1,-1 0,-0.5527 -0.4478,-1 -1,-1 z M 24.300817,45.804787 c -1.103,0 -2,-0.897 -2,-2 0,-1.103 0.897,-2 2,-2 1.103,0 2,0.897 2,2 0,1.103 -0.897,2 -2,2 z m -0.204292,25.964246 c -1.103,0 -2,-0.8975 -2,-2 0,-1.1025 0.897,-2 2,-2 1.103,0 2,0.8975 2,2 0,1.1025 -0.897,2 -2,2 z"/>
+          </g>
+        </svg>
+      </div>
+    </div>
+    <div class="splash-name">alan barzilay<span class="dot">.</span></div>
+    <div class="splash-sub">developer · researcher · maker</div>
+  </div>
+</div>
