+++ src/components/landing/TunnelStage.astro
@@ -1,0 +1,7 @@
+---
+---
+<!-- ====================== TUNNEL ====================== -->
+<canvas id="tunnel-canvas"></canvas>
+<div class="layer" id="tunnel-ui">
+  <div id="cardStage"></div>
+</div>
